requirements.txt
aiofiles==0.4.0
alembic==1.1.0
atomicwrites==1.3.0
attrs==19.1.0
autocomplete==0.0.104
bcrypt==3.1.7
blinker==1.4
bottle==0.12.17
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
coverage==4.5.4
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
flask-swagger-ui==3.20.9
Flask-WTF==0.14.2
gunicorn==19.9.0
h11==0.9.0
h2==3.1.1
hpack==3.0.0
Hypercorn==0.8.2
hyperframe==5.2.0
idna==2.8
importlib-metadata==0.20
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==2.20.2
more-itertools==7.2.0
multidict==4.5.2
packaging==19.1
pluggy==0.12.0
priority==1.3.0
psycopg2==2.8.3
py==1.8.0
pycparser==2.19
PyJWT==1.7.1
pyparsing==2.4.2
pytest==5.1.2
pytest-cov==2.7.1
python-coveralls==2.9.3
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
PyYAML==5.1.2
Quart==0.10.0
Quart-CORS==0.2.0
requests==2.22.0
six==1.12.0
sortedcontainers==2.1.0
SQLAlchemy==1.3.8
toml==0.10.0
typing-extensions==3.7.4
urllib3==1.25.3
wcwidth==0.1.7
Werkzeug==0.15.5
wsproto==0.15.0
WTForms==2.2.1
zipp==0.6.0
